`forge fmt`

+++ contracts/LlamaLendFactory.sol
@@ -1,7 +1,7 @@
 //SPDX-License-Identifier: None
 pragma solidity ^0.8.0;
 
-import './LendingPool.sol';
+import "./LendingPool.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 import "@openzeppelin/contracts/proxy/Clones.sol";
 
@@ -9,7 +9,6 @@ import "@openzeppelin/contracts/proxy/Clones.sol";
 /// @author 0xngmi
 /// @dev Factory for LendingPool; batches shutdowns and loan repayments
 contract LlamaLendFactory is Ownable {
-
     /*//////////////////////////////////////////////////////////////
                                LIBRARIES
     //////////////////////////////////////////////////////////////*/
@@ -51,20 +50,36 @@ contract LlamaLendFactory is Ownable {
     //////////////////////////////////////////////////////////////*/
 
     function createPool(
-        address _oracle, uint _maxPrice, address _nftContract, 
-        uint _maxDailyBorrows, string memory _name, string memory _symbol,
-        uint96 _maxLoanLength, LendingPool.Interests calldata interests
+        address _oracle,
+        uint256 _maxPrice,
+        address _nftContract,
+        uint256 _maxDailyBorrows,
+        string memory _name,
+        string memory _symbol,
+        uint96 _maxLoanLength,
+        LendingPool.Interests calldata interests
     ) external returns (LendingPool pool) {
         require(_maxLoanLength < 1e18, "maxLoanLength too big"); // 31bn years, makes sure that reverts cant be forced through this
         pool = LendingPool(address(implementation).clone());
-        pool.initialize(_oracle, _maxPrice, _maxDailyBorrows, _name, _symbol, interests, msg.sender, _nftContract, address(this), _maxLoanLength);
+        pool.initialize(
+            _oracle,
+            _maxPrice,
+            _maxDailyBorrows,
+            _name,
+            _symbol,
+            interests,
+            msg.sender,
+            _nftContract,
+            address(this),
+            _maxLoanLength
+        );
         emit PoolCreated(_nftContract, msg.sender, address(pool));
     }
-    
+
     function repay(LoanRepayment[] calldata loansToRepay) external payable {
-        uint length = loansToRepay.length;
-        uint i = 0;
-        while(i<length){
+        uint256 length = loansToRepay.length;
+        uint256 i = 0;
+        while (i < length) {
             LendingPool(loansToRepay[i].pool).repay{value: address(this).balance}(loansToRepay[i].loans, msg.sender);
             unchecked {
                 i++;
@@ -78,7 +93,7 @@ contract LlamaLendFactory is Ownable {
     //////////////////////////////////////////////////////////////*/
 
     function emergencyShutdown(address[] calldata pools) external onlyOwner {
-        for(uint i = 0; i < pools.length; i++){
+        for (uint256 i = 0; i < pools.length; i++) {
             LendingPool(pools[i]).emergencyShutdown();
         }
     }
